requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.19.0
albumentations==1.3.1
alembic==1.11.1
amqp==5.1.1
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blessed==1.20.0
blinker==1.6.2
blosc2==2.0.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
cmake==3.26.4
colorama==0.4.6
colorlog==6.7.0
comm==0.1.3
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.0.7
croniter==1.3.15
cryptography==41.0.1
cupy==12.3.0
cycler==0.11.0
Cython==3.0.8
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databricks-cli==0.17.7
dateutils==0.6.12
debugpy==1.6.7
decorator==5.1.1
deepdiff==6.3.0
defusedxml==0.7.1
deprecation==2.1.0
dictdiffer==0.9.0
dill==0.3.6
diskcache==5.6.1
distro==1.8.0
dlib==19.24.2
docker==6.1.3
docker-pycreds==0.4.0
dpath==2.1.6
dulwich==0.21.5
dvc==3.2.1
dvc-data==2.3.0
dvc-http==2.30.2
dvc-objects==0.23.0
dvc-render==0.5.3
dvc-studio-client==0.10.0
dvc-task==0.3.0
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.88.0
fastjsonschema==2.18.0
fastrlock==0.8.2
ffmpeg==1.4
filelock==3.12.0
filterpy==1.4.5
Flask==2.2.5
flatbuffers==23.5.26
flatten-dict==0.4.2
flufl-lock==8.0.1
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
funcy==2.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
hydra-core==1.3.2
hydra-joblib-launcher==1.2.0
idna==3.4
imageio==2.31.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
ipykernel==6.23.3
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
iterative-telemetry==0.0.8
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==5.3.1
jupyter_packaging==0.12.3
jupyter_server==2.7.3
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
kiwisolver==1.4.4
kombu==5.3.1
lap==0.4.0
lazy_loader==0.3
lightning==2.2.0.post0
lightning-cloud==0.5.36
lightning-utilities==0.8.0
lit==16.0.5.post0
loguru==0.7.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.2
mistune==3.0.1
mlflow==2.10.2
mpi4py==3.1.4
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.14
nanotime==0.5.2
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.5
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.23.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open3d==0.17.0
opencv-contrib-python==4.8.0.74
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
ordered-set==4.1.0
orjson==3.9.1
overrides==7.4.0
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
plotly==5.16.1
pox==0.3.2
ppft==1.7.6.6
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
pygit2==1.12.1
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.7.0
pyparsing==3.0.9
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==2.0.3
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
querystring-parser==1.2.4
readchar==4.0.5
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.1
rpds-py==0.10.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
scmrepo==1.0.4
seaborn==0.12.2
Send2Trash==1.8.2
sentry-sdk==1.26.0
setproctitle==1.3.2
shortuuid==1.0.11
shtab==1.6.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.4.1
SQLAlchemy==2.0.15
sqlparse==0.4.4
sqltrie==0.7.0
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tables==3.8.0
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.10
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
torch==2.2.0
torch-spatiotemporal==0.9.4
torch_geometric==2.4.0
torch_scatter==2.1.2
torch_sparse==0.6.18
torchmetrics==1.1.2
torchvision==0.17.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.2.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
vine==5.0.0
voluptuous==0.13.1
wandb==0.16.0
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.8
y-py==0.6.0
yarl==1.9.2
ypy-websocket==0.8.4
zc.lockfile==3.0.post1
zipp==3.15.0
